update tests for actions
